Implemented an AbstractRequestBase class to include methods common to all request types.
